refactor: moved empty square component to icons directory
