+++ src/components/Icons/EmptySquare/EmptySquare.tsx
