+++ src/components/Icons/EmptySquare/index.ts
